+++ src/main/scala/de/m7w3/signal/QRCodeGenerator.scala
@@ -1,0 +1,10 @@
+package de.m7w3.signal
+
+import net.glxn.qrgen.javase.QRCode
+import java.io.ByteArrayOutputStream
+
+object QRCodeGenerator {
+  def generate(getUrl: () => String): ByteArrayOutputStream = {
+    QRCode.from(getUrl()).stream
+  }
+}
